fix test_multiprocess_dataloader_base timeout. test=develop
